// apps/knowledge-base/app/(app)/[locale]/(dashboard)/dashboard/administrator/project-persons/_lib/upsert-project-person.schema.ts
import * as v from "valibot";

export const UpsertProjectPersonActionInputSchema = v.object({
	id: v.optional(v.pipe(v.string(), v.uuid())),
	projectDocumentId: v.pipe(v.string(), v.uuid()),
	personDocumentId: v.pipe(v.string(), v.uuid()),
	roleId: v.pipe(v.string(), v.uuid()),
	duration: v.optional(
		v.object({
			start: v.pipe(v.string(), v.isoDate(), v.toDate()),
			end: v.optional(v.pipe(v.string(), v.isoDate(), v.toDate())),
		}),
	),
});